Add Github Action to push Docker Images
After a release is published, the corresponding Github event will
trigger this workflow to build and push the Docker images for the VMware
Event Router as per the Makefile target `make release`.

Github secrets (encrypted) for `DOCKER_USER` and `DOCKER_SECRET` are
used to authenticate. These secrets are not exposed through the UI and
not persisted after the workflow terminates as per the Github Actions
documentation.

Signed-off-by: Michael Gasch <mgasch@vmware.com>