Implement at-least-once delivery via checkpointing
In order to support at-least-once delivery semantics to guard against
crashes of the event router and thus missing events this feature adds
checkpointing logic to the vCenter event provider.

Checkpointing can be enabled via a configuration flag for the vCenter
event provider and optionally allows to specify a custom path for the
checkpoint files.

Signed-off-by: Michael Gasch <mgasch@vmware.com>