Add invocation details to metrics
In order to easily indicate success/failure for processor invocations, a concrete struct `InvocationDetails` is added to the metrics. It uses *T semantics to persist changes. It is important to initialize the `Invocations` map correctly and check for nil values due to *T. Callers must use the `Success` and `Failure` methods. `EventStats` is not thread-safe.

Signed-off-by: Michael Gasch <mgasch@vmware.com>