Add retries to OpenFaaS processor
To simplify the code base the OpenFaaS processor code is moved to its own (sub) package under the `processor` tree.

Retries were added, based on a forked change in the controller SDK.

Changes were made to the methods to comply with the updated `processor` interface and changes to `map[string]*metrics.InvocationDetails`.

Signed-off-by: Michael Gasch <mgasch@vmware.com>