Update install-knative.md
Added a new note below the Event Processor selection.  This is a detail, but when I was going through deployment, it was not clear that the Knative configuration section was only needed for external processors/brokers.  The note may be a little too wordy, but I think it is an distinction that we need to make in the deployment doc.