feat: Add vsphereapiversion to CE context
Since vSphere events are not versioned and (somewhat loosely) tied to a
particular vSphere API version, this change introduces a new CloudEvents
context attribute `vsphereapiversion` reflecting the vSphere API version
of a particular CloudEvent produced by the `vcenter` event producer.

BREAKING: This change sets the `timestamp` in the CloudEvent to the
`CreatedTime` as set by vCenter in a vSphere event instead of
`time.Now()`.

Closes: #439
Signed-off-by: Michael Gasch <mgasch@vmware.com>