Added aws_iam_role AuthMethod for AWS EventBridge
Adjusted the config and aws_event_bridge processor to support the aws_iam_role AuthMethod. When this aws_iam_role is used the AWS SDK session for the AWS EventBridge processor will load configuration and credentials from the environment, configuration files, and other credential sources. The purpose of this is to enable short-term credentials generated from assumed role on the executing pod or EC2 machine to be used.

Signed-off-by: Adrian Begg <adriabeg@amazon.com>